Use akka-seed-1 if it is available, akka-seed-0 otherwise
For new clusters, the first node needs to talk to itself as its seed node.
For existing clusters, an existing member needs to be contacted.

In order to avoid akka-seed-0 starting new clusters every time it restarts,
nodes contact akka-seed-1 if a cluster already exists there.
Otherwise, a non-existing cluster is assumed and akka-seed-0 becomes
the seed node.